updated the comment service

+++ skillhub/src/main/java/com/skillhub/skillhub/service/CommentService.java
@@ -1,6 +1,5 @@
 package com.skillhub.skillhub.service;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,35 +15,67 @@ public class CommentService {
     @Autowired
     private CommentRepository commentRepository;
 
-    public Comment addComment(String postId, Comment comment) {
-        comment.setPostId(postId);
+    public Comment createComment(Comment comment) {
         return commentRepository.save(comment);
     }
 
+    public Comment createReply(String parentCommentId, Comment reply) {
+        Optional<Comment> parentComment = commentRepository.findById(parentCommentId);
+        if (parentComment.isPresent()) {
+            reply.setParentCommentId(parentCommentId);
+            reply.setPostId(parentComment.get().getPostId());
+            Comment savedReply = commentRepository.save(reply);
+            
+            // Add reply to parent's replies list
+            Comment parent = parentComment.get();
+            parent.getReplies().add(savedReply);
+            commentRepository.save(parent);
+            
+            return savedReply;
+        }
+        throw new RuntimeException("Parent comment not found");
+    }
+
     public List<Comment> getCommentsByPostId(String postId) {
-        return commentRepository.findAllByPostId(postId);
+        return commentRepository.findByPostId(postId);
     }
 
-    public Optional<Comment> getCommentById(String postId, String id) {
+    public List<Comment> getTopLevelComments(String postId) {
+        return commentRepository.findTopLevelComments(postId);
+    }
+
+    public List<Comment> getReplies(String parentCommentId) {
+        return commentRepository.findByParentCommentId(parentCommentId);
+    }
+
+    public Comment updateComment(String id, Comment updatedComment) {
+        Optional<Comment> existingComment = commentRepository.findById(id);
+        if (existingComment.isPresent()) {
+            Comment comment = existingComment.get();
+            comment.setText(updatedComment.getText());
+            comment.setEdited(true);
+            return commentRepository.save(comment);
+        }
+        throw new RuntimeException("Comment not found");
+    }
+
+    public void deleteComment(String id) {
         Optional<Comment> comment = commentRepository.findById(id);
-        return comment.filter(c -> c.getPostId().equals(postId));
+        if (comment.isPresent()) {
+            // If it's a reply, remove it from parent's replies
+            if (comment.get().getParentCommentId() != null) {
+                Optional<Comment> parentComment = commentRepository.findById(comment.get().getParentCommentId());
+                if (parentComment.isPresent()) {
+                    Comment parent = parentComment.get();
+                    parent.getReplies().removeIf(reply -> reply.getId().equals(id));
+                    commentRepository.save(parent);
+                }
+            }
+            commentRepository.deleteById(id);
+        }
     }
 
-    public Comment updateComment(String postId, String id, Comment updatedComment) {
-        Comment existingComment = commentRepository.findById(id)
-                .filter(c -> c.getPostId().equals(postId))
-                .orElseThrow(() -> new RuntimeException("Comment not found with id: " + id + " and postId: " + postId));
-
-        existingComment.setText(updatedComment.getText());
-        existingComment.setAuthor(updatedComment.getAuthor());
-        return commentRepository.save(existingComment);
-    }
-
-    public void deleteComment(String postId, String id) {
-        commentRepository.findById(id)
-                .filter(c -> c.getPostId().equals(postId))
-                .orElseThrow(() -> new RuntimeException("Comment not found with id: " + id + " and postId: " + postId));
-
-        commentRepository.deleteById(id);
+    public void deleteCommentsByPostId(String postId) {
+        commentRepository.deleteByPostId(postId);
     }
 }
